updated Display of Profile and added use case ChangeBalanceVisibility
