+++ profile/src/main/java/rdx/works/profile/data/model/apppreferences/Display.kt
@@ -4,7 +4,7 @@ import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
 
 @Serializable
-class Display private constructor(
+class Display(
     @SerialName("fiatCurrencyPriceTarget")
     val fiatCurrencyPriceTarget: String,
     @SerialName("isCurrencyAmountVisible")
