+++ profile/src/main/java/rdx/works/profile/domain/GetProfileUseCase.kt
@@ -183,6 +183,9 @@ val GetProfileUseCase.security
 val GetProfileUseCase.p2pLinks
     get() = invoke().map { it.appPreferences.p2pLinks }
 
+val GetProfileUseCase.isBalanceVisible
+    get() = invoke().map { it.appPreferences.display.isCurrencyAmountVisible }
+
 /**
  * Default deposit guarantee
  */
